+++ src/hwenc_amf/amf_video_encoder.cpp
@@ -541,9 +541,8 @@ AMF_RESULT AMFVideoEncoderImpl::ProcessBuffer(amf::AMFBufferPtr buffer,
   webrtc::EncodedImageCallback::Result result =
       callback_->OnEncodedImage(encoded_image_, &codec_specific);
   if (result.error != webrtc::EncodedImageCallback::Result::OK) {
-    RTC_LOG(LS_ERROR) << __FUNCTION__
-                      << " OnEncodedImage failed error:" << result.error;
-    return AMF_FAIL;
+    RTC_LOG(LS_WARNING) << __FUNCTION__
+                        << " OnEncodedImage failed error:" << result.error;
   }
   bitrate_adjuster_.Update(size);